refactor: extract mod `raft` into crate member
