+++ crates/raft/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+    name = "raft"
+    version = "0.1.0"
+    edition = "2021"
+
+[dependencies]
+    timer.workspace = true
+    anyhow = "1.0.86"
+    futures = "0.3.30"
+    log = {version = "0.4.22", features = [
+        "kv",
+    ]}
+    prost = "0.13.1"
+    rand = "0.8.5"
+    serde = {version = "1.0.207", features = [
+        "derive",
+    ]}
+    serde_json = "1.0.124"
+    structured-logger = "1.0.3"
+    tokio = {version = "1.39.2", features = [
+        "macros",
+        "rt-multi-thread",
+    ]}
+    tonic = "0.12.1"
+
+[build-dependencies]
+    tonic-build = "0.11.0"
