+++ crates/raft/src/state/candidate.rs
@@ -5,7 +5,7 @@ use super::{
     RequestVoteArgs, RequestVoteReply,
 };
 use super::{PeerID, RaftContext, Role, State, Term};
-use crate::raft::config;
+use crate::config;
 use futures::future;
 use log::{debug, error, info};
 use serde::Serialize;
