+++ crates/raft/tests/raft_test.rs
@@ -1,6 +1,6 @@
-use radis::raft::config::Config;
-use radis::raft::state;
-use radis::raft::RaftService;
+use raft::config::Config;
+use raft::state;
+use raft::RaftService;
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::mpsc;
